insurance/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.4
numpy==1.20.1
pandas==1.2.3
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
